requirements.txt
anyio==3.6.2
asgiref==3.5.2
asttokens==2.0.8
asyncio==3.4.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2021.10.8
chardet==5.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
dataclasses==0.6
DateTime==4.5
debugpy==1.6.3
decorator==5.1.1
deptry==0.6.3
dnspython==2.2.1
entrypoints==0.4
exceptiongroup==1.0.0
executing==1.0.0
fastapi
python-dotenv
uvicorn[standard]>=0.17.6
python-multipart>=0.0.5
pefile>=2023.2.7
future==0.18.2
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httplib2==0.21.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.15.2
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
lxml==4.9.1
matplotlib-inline==0.1.6
motor==3.0.0
# mysqlclient==2.1.1   # Bật nếu bạn cần MySQL C API
# mysql-connector-python==8.0.31  # Bật nếu cần MySQL Connector (lưu ý conflict)
nest-asyncio==1.5.5
numpy>=2.0.0
oauth2client==4.1.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.4
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.31
protobuf==4.21.12
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
PyDrive==1.3.1
Pygments==2.13.0
PyJWT==2.6.0
pymongo==4.1.1
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.0
starlette==0.19.1
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.17.6
wcwidth==0.2.5
yara-python==4.2.3
zope.interface==5.4.0
Flask>=2.2.0
Werkzeug>=2.2.0
aiomysql==0.2.0
jinja2>=3.1.2
